+++ src/FAPOFX_echo.c
@@ -148,7 +148,7 @@ void FAPOFXEcho_Process(
 	const FAPOProcessBufferParameters* pInputProcessParameters,
 	uint32_t OutputProcessParameterCount,
 	FAPOProcessBufferParameters* pOutputProcessParameters,
-	uint8_t IsEnabled
+	int32_t IsEnabled
 ) {
 	FAPOBase_BeginProcess(&fapo->base);
