+++ src/FAPOFX_eq.c
@@ -148,7 +148,7 @@ void FAPOFXEQ_Process(
 	const FAPOProcessBufferParameters* pInputProcessParameters,
 	uint32_t OutputProcessParameterCount,
 	FAPOProcessBufferParameters* pOutputProcessParameters,
-	uint8_t IsEnabled
+	int32_t IsEnabled
 ) {
 	FAPOBase_BeginProcess(&fapo->base);
